+++ tests/Regressors/SVRTest.php
@@ -17,7 +17,7 @@ class SVRTest extends TestCase
 {
     const TRAIN_SIZE = 300;
     const TEST_SIZE = 5;
-    const TOLERANCE = 5;
+    const TOLERANCE = 10;
 
     protected $generator;